continuing work on settings
settings are now in development. UI is partial. moved some of the fxml
files to a different package. this will help differentiate fxml files
for different programs but allow for similar names to be used. Settings
are now being worked on although they will take a little while to get
finished. changed how you add a user, now only takes two params but one
param is an object. pass in -1 as default value for userId on creating a
user if adding to database. userId has no use until later when adding
packages.