working to finish some things
settings page has a few differences. namely the addition of a cancel
button. upon saving settings, user will be prompted wether or not to
restart the program (little trickery there). javaFX jar library is
included (makes it compatible with jre7). necessary for running the
program on versions of java prior to java 8.