lot more work here
working on finishing up settings page. small bug found in openpage
controller where if table was selected two key events would be
dispatched. not a big deal for refreshing but for logout key event it
was causing duplicate messages to appear. added new methods to
databasemanager and subsequently to sqlitemanager. these methods pertain
to new account management abilities in settings. also added a new
setting type called "AUFREQ" which is the auto update frequency. this
will be implemented at some point in the future by using a seperate
thread and a timer for it. in scan page, autofill will match the case of
the person pulled from the database. this will alleviate issues when
searching for packages(lowercase sorts below ALL uppercase variants).