serious changes
small change to OpenPageController. implemented the Auto Update Feature.
this is based on the slider in the general tab in settings. in
SQLiteManager, been refactored. old db schema was being used for
compatibility reasons but this approach is abandoned. moving forward
with new schema which will match that of the MYSQL version, except for
an added column for hash password from PHP side.