working to set stop as default
currently on otherclient loading the default is to load ALL packages.
changing it so that a specific stop can be set as the default and then
if no stop is set as default will load ALL packages. functionality added
to various database managers. have to change Stop class a bit to allow
for it.