added a feature that was requested
it was requested that when printing stops that were for students would
be sorted by box number as opposed to the default which was last name,
first name. this has been implemented in a separate method so as not to
interfere with normal database retrieval for other parts of the program.
tested in sqlite to be working. additional tests will be done on other
database types.