fixed bugs with launching other jar files
on windows using 'java' as command in ProcessBuilder will not work.
linux does work though. in Windows you can use Desktop.open(file) and
that works. but in linux that will not work as it changes the working
directory and we rely on that for finding configuration files.
therefore, i wrapped os.name property to determine which way to work
with