all v3 work
added some new functionality for demo purposes. also changing how the
backend works. looking to remove the java 1.7 ability from the system.
Added new objects to correlate with some of teh new background changes.
basically all this will be used in a different project which will be a
full ground up rewrite of the system as V2 is having many issues as of
late